+++ packages/reakit/src/Dialog/DialogState.ts
@@ -20,6 +20,10 @@ export type DialogState = DisclosureState & {
    * by default.
    */
   modal: boolean;
+  /**
+   * @private
+   */
+  unstable_disclosureRef: React.MutableRefObject<HTMLElement | null>;
 };
 
 export type DialogActions = DisclosureActions & {
@@ -34,12 +38,7 @@ export type DialogInitialState = DisclosureInitialState &
 
 export type DialogStateReturn = DisclosureStateReturn &
   DialogState &
-  DialogActions & {
-    /**
-     * @private
-     */
-    unstable_modal: boolean;
-  };
+  DialogActions;
 
 export function useDialogState(
   initialState: SealedInitialState<DialogInitialState> = {}
@@ -48,14 +47,15 @@ export function useDialogState(
     initialState
   );
 
+  const disclosure = useDisclosureState(sealed);
   const [modal, setModal] = React.useState(initialModal);
-  const hidden = useDisclosureState(sealed);
+  const disclosureRef = React.useRef<HTMLElement | null>(null);
 
   return {
+    ...disclosure,
     modal,
     setModal,
-    unstable_modal: modal,
-    ...hidden,
+    unstable_disclosureRef: disclosureRef,
   };
 }
 
@@ -63,7 +63,7 @@ const keys: Array<keyof DialogStateReturn> = [
   ...useDisclosureState.__keys,
   "modal",
   "setModal",
-  "unstable_modal",
+  "unstable_disclosureRef",
 ];
 
 useDialogState.__keys = keys;