+++ packages/reakit/src/Dialog/__tests__/index-test.tsx
@@ -20,20 +20,6 @@ test("clicking on disclosure opens the dialog", () => {
   expect(dialog).toBeVisible();
 });
 
-test("passing modal to Dialog should warn", () => {
-  const Test = () => {
-    const dialog = useDialogState();
-    return (
-      <>
-        <DialogDisclosure {...dialog}>disclosure</DialogDisclosure>
-        <Dialog {...dialog} tabIndex={0} modal={false} aria-label="dialog" />
-      </>
-    );
-  };
-  render(<Test />);
-  expect(console).toHaveWarned();
-});
-
 test("focus the first tabbable element when dialog opens", () => {
   const Test = () => {
     const dialog = useDialogState();
@@ -605,30 +591,6 @@ test("focusing disclosure does not close the non-modal dialog", () => {
   expect(dialog).toBeVisible();
 });
 
-test("focusing any of multiple disclosures does not close the non-modal dialog", () => {
-  const Test = () => {
-    const dialog = useDialogState({ visible: true, modal: false });
-    return (
-      <>
-        <DialogDisclosure {...dialog}>disclosure1</DialogDisclosure>
-        <DialogDisclosure {...dialog}>disclosure2</DialogDisclosure>
-        <Dialog {...dialog} aria-label="dialog" />
-      </>
-    );
-  };
-  const { getByText, getByLabelText } = render(<Test />);
-  const disclosure1 = getByText("disclosure1");
-  const disclosure2 = getByText("disclosure2");
-  const dialog = getByLabelText("dialog");
-  expect(dialog).toBeVisible();
-  focus(disclosure1);
-  expect(disclosure1).toHaveFocus();
-  expect(dialog).toBeVisible();
-  focus(disclosure2);
-  expect(disclosure2).toHaveFocus();
-  expect(dialog).toBeVisible();
-});
-
 test("focus disclosure when dialog closes", () => {
   const Test = () => {
     const dialog = useDialogState();
@@ -794,39 +756,6 @@ test("focus the first tabbable element when nested dialog opens", () => {
   expect(button1).toHaveFocus();
   click(disclosure2);
   expect(button2).toHaveFocus();
-});
-
-test("focus disclosure in dialog when nested dialog closes", () => {
-  const Test = () => {
-    const dialog = useDialogState();
-    const dialog2 = useDialogState();
-    return (
-      <>
-        <DialogDisclosure {...dialog}>disclosure1</DialogDisclosure>
-        <Dialog {...dialog} tabIndex={undefined} aria-label="dialog1">
-          <button>button1</button>
-          <DialogDisclosure {...dialog2}>disclosure2</DialogDisclosure>
-          <Dialog {...dialog2} aria-label="dialog2">
-            <button>button2</button>
-            <button>button3</button>
-          </Dialog>
-        </Dialog>
-      </>
-    );
-  };
-  const { getByText, getByLabelText } = render(<Test />);
-  const disclosure1 = getByText("disclosure1");
-  const dialog1 = getByLabelText("dialog1");
-  const disclosure2 = getByText("disclosure2");
-  const dialog2 = getByLabelText("dialog2");
-  const button2 = getByText("button2");
-  click(disclosure1);
-  focus(disclosure2);
-  click(disclosure2);
-  expect(button2).toHaveFocus();
-  click(dialog1);
-  expect(dialog2).not.toBeVisible();
-  expect(disclosure2).toHaveFocus();
 });
 
 test("focus is trapped within the nested dialog", () => {
@@ -1401,4 +1330,25 @@ test("nested modal dialog with backdrop markup", () => {
       />
     </body>
   `);
+});
+
+test("should render the default tabIndex when none is specified", () => {
+  function Test() {
+    const dialog = useDialogState();
+    return <Dialog {...dialog} aria-label="dialog" />;
+  }
+
+  const { getByLabelText } = render(<Test />);
+  expect(getByLabelText("dialog")).toHaveAttribute("tabIndex", "-1");
+});
+
+// See https://github.com/reakit/reakit/issues/636
+test("passing undefined to tabIndex should render the default tabIndex of -1", () => {
+  function Test() {
+    const dialog = useDialogState();
+    return <Dialog {...dialog} aria-label="dialog" tabIndex={undefined} />;
+  }
+
+  const { getByLabelText } = render(<Test />);
+  expect(getByLabelText("dialog")).toHaveAttribute("tabIndex", "-1");
 });