+++ packages/reakit/src/Tabbable/Tabbable.ts
@@ -2,12 +2,11 @@ import * as React from "react";
 import { createComponent } from "reakit-system/createComponent";
 import { createHook } from "reakit-system/createHook";
 import { useForkRef } from "reakit-utils/useForkRef";
+import { useIsomorphicEffect } from "reakit-utils/useIsomorphicEffect";
 import { useLiveRef } from "reakit-utils/useLiveRef";
-import { isFocusable } from "reakit-utils/tabbable";
-import { hasFocusWithin } from "reakit-utils/hasFocusWithin";
-import { isButton } from "reakit-utils/isButton";
 import { warning } from "reakit-warning";
 import { BoxOptions, BoxHTMLProps, useBox } from "../Box/Box";
+import { useFocusOnMouseDown } from "./__utils/useFocusOnMouseDown";
 
 export type TabbableOptions = BoxOptions & {
   /**
@@ -28,6 +27,16 @@ export type TabbableHTMLProps = BoxHTMLProps & {
 
 export type TabbableProps = TabbableOptions & TabbableHTMLProps;
 
+function isUserAgent(string: string) {
+  if (typeof window === "undefined") return false;
+  return window.navigator.userAgent.indexOf(string) !== -1;
+}
+
+const isSafariOrFirefoxOnMac =
+  isUserAgent("Mac") &&
+  !isUserAgent("Chrome") &&
+  (isUserAgent("Safari") || isUserAgent("Firefox"));
+
 function isNativeTabbable(element: Element) {
   return (
     element.tagName === "BUTTON" ||
@@ -39,15 +48,6 @@ function isNativeTabbable(element: Element) {
     element.tagName === "VIDEO"
   );
 }
-
-// https://twitter.com/diegohaz/status/1176998102139572225
-function isUserAgent(string: string) {
-  if (typeof window === "undefined") return false;
-  return window.navigator.userAgent.indexOf(string) !== -1;
-}
-
-const isSafariOrFirefoxOnMac =
-  isUserAgent("Mac") && (isUserAgent("Safari") || isUserAgent("Firefox"));
 
 export const useTabbable = createHook<TabbableOptions, TabbableHTMLProps>({
   name: "Tabbable",
@@ -78,8 +78,11 @@ export const useTabbable = createHook<TabbableOptions, TabbableHTMLProps>({
     const style = options.disabled
       ? { pointerEvents: "none" as const, ...htmlStyle }
       : htmlStyle;
+    const onMouseDownFocus = isSafariOrFirefoxOnMac
+      ? useFocusOnMouseDown()
+      : undefined;
 
-    React.useEffect(() => {
+    useIsomorphicEffect(() => {
       const tabbable = ref.current;
       if (!tabbable) {
         warning(
@@ -113,23 +116,10 @@ export const useTabbable = createHook<TabbableOptions, TabbableHTMLProps>({
           event.preventDefault();
           return;
         }
-
+        onMouseDownFocus?.(event);
         onMouseDownRef.current?.(event);
-        if (event.defaultPrevented) return;
-
-        const self = event.currentTarget;
-        const target = event.target as HTMLElement;
-
-        if (isSafariOrFirefoxOnMac && isButton(self) && self.contains(target)) {
-          event.preventDefault();
-          const isFocusControl =
-            isFocusable(target) || target.tagName === "LABEL";
-          if (!hasFocusWithin(self) || self === target || !isFocusControl) {
-            self.focus();
-          }
-        }
       },
-      [options.disabled]
+      [options.disabled, onMouseDownFocus]
     );
 
     return {