+++ packages/website/src/components/SEO.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Helmet from "react-helmet";
+import { Helmet } from "react-helmet";
 import { graphql, useStaticQuery } from "gatsby";
 import thumbnail from "../images/thumbnail.png";