+++ packages/reakit-system-bootstrap/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.1](https://github.com/reakit/reakit/tree/master/packages/reakit-system-bootstrap/compare/reakit-system-bootstrap@0.11.0...reakit-system-bootstrap@0.11.1) (2020-05-12)
+
+**Note:** Version bump only for package reakit-system-bootstrap
+
+
+
+
+
+# [0.11.0](https://github.com/reakit/reakit/tree/master/packages/reakit-system-bootstrap/compare/reakit-system-bootstrap@0.10.0...reakit-system-bootstrap@0.11.0) (2020-04-29)
+
+**Note:** Version bump only for package reakit-system-bootstrap
+
+
+
+
+
 # [0.10.0](https://github.com/reakit/reakit/tree/master/packages/reakit-system-bootstrap/compare/reakit-system-bootstrap@0.9.0...reakit-system-bootstrap@0.10.0) (2020-04-20)