+++ packages/reakit-warning/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/reakit/reakit/tree/master/packages/reakit-warning/compare/reakit-warning@0.2.0...reakit-warning@0.3.0) (2020-04-29)
+
+**Note:** Version bump only for package reakit-warning
+
+
+
+
+
 # [0.2.0](https://github.com/reakit/reakit/tree/master/packages/reakit-warning/compare/reakit-warning@0.1.0...reakit-warning@0.2.0) (2020-04-20)
 
 **Note:** Version bump only for package reakit-warning